Moving get_fattime() out of fatfs
