--- src/fatfs/fattime.h
@@ -1,8 +1,0 @@
-#ifndef FATTIME_H_
-#define FATTIME_H_
-
-#include "ff.h"
-
-DWORD get_fattime (void);
-
-#endif
