+++ src/fatfs/ff.c
@@ -22,7 +22,6 @@
 #include <string.h>
 #include "ff.h"			/* Declarations of FatFs API */
 #include "diskio.h"		/* Declarations of device I/O functions */
-#include "fattime.h"
 #include "pico/platform.h"
