+++ src/ria/api/rtc.c
@@ -66,3 +66,20 @@ err_param:
     xstack_ptr = XSTACK_SIZE;
     api_return_errno_axsreg_zxstack(1, -1);
 }
+
+DWORD get_fattime (void)
+{
+	DWORD res;
+	datetime_t datetime;
+
+	rtc_get_datetime(&datetime);
+	
+	res =  (((DWORD)datetime.year - 1980) << 25)
+			| ((DWORD)datetime.month << 21)
+			| ((DWORD)datetime.day << 16)
+			| (WORD)(datetime.hour << 11)
+			| (WORD)(datetime.min << 5)
+			| (WORD)(datetime.sec >> 1);
+
+	return res;
+}
